+++ crates/idxdb-store/src/lib.rs
@@ -148,7 +148,8 @@ impl IdxdbStore {
     }
 }
 
-#[async_trait::async_trait(?Send)]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 impl Store for IdxdbStore {
     fn identifier(&self) -> &str {
         &self.database_id